new refactoring, add transaction to create order

+++ app/Http/Services/ApiService/OrdersService.php
@@ -1,31 +1,46 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Services\ApiService;
 
 use App\Http\Middleware\Authenticate;
 use App\Http\Requests\Order\CreateOrder;
 use App\Http\Resources\Order\OrderResource;
 use App\Models\Order;
-use App\Models\OrderItems;
 use App\Models\Payment;
 use App\Models\Status;
 use App\Models\User;
 use Carbon\Carbon;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Resources\Json\ResourceCollection;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\DB;
 
 class OrdersService
 {
 
     /**
-     * @return JsonResponse
-     * Получить заказы юзера
+     * Получить все заказы
      */
-    public function getUserOrders(): JsonResponse
+    public function getOrders(): ResourceCollection
     {
+//        return Cache::remember(Order::CACHE_NAME, Order::CACHE_TIME, function () {
         $user = Auth::user();
-        return response()->json(OrderResource::collection($user->orders));
+        return OrderResource::collection($user->orders);
+//        });
+    }
+
+    /**
+     * @param Order $order
+     * @return OrderResource
+     * Получить заказ
+     */
+    public function getOrder(Order $order): OrderResource
+    {
+        return new OrderResource($order);
     }
 
     /**
@@ -36,52 +51,65 @@ class OrdersService
     public function createNewOrder(CreateOrder $request): JsonResponse
     {
         $user = Auth::user();
+        $created = false;
 
-        /** @var User $user */
+        /**
+         * @var User|Authenticatable $user
+         */
 
         if ($user?->cart) {
-            $order = new Order();
-            $order->fill([
-                'user_id' => Auth::id(),
-                'status_id' => Status::JUST_CREATED,
-                'payment_id' => $request->validated('payment_id'),
-                'delivery_time' => Carbon::parse($request->validated('delivery_time'))->format('d-m-y H:i:s'),
-                'address' => $request->validated('address'),
-                'phone' => $request->validated('phone'),
-                'comment' => $request->validated('comment'),
-                'closed' => false,
-                'total_sum' => $user->cart->items->sum('price'),
-            ]);
-            $order->save();
+            $created = DB::transaction(function () use ($request, $user) {
+                $order = new Order();
+                try {
+                    $order->fill([
+                        'user_id' => Auth::id(),
+                        'status_id' => Status::JUST_CREATED,
+                        'payment_id' => $request->validated('payment_id'),
+                        'delivery_time' => Carbon::parse($request->validated('delivery_time'))->format('d-m-y H:i:s'),
+                        'address' => $request->validated('address'),
+                        'phone' => $request->validated('phone'),
+                        'comment' => $request->validated('comment'),
+                        'closed' => false,
+                        'total_sum' => $user->cart->items->sum('price'),
+                    ]);
+                    $order->save();
 
-            foreach ($user->cart->items as $cartItem) {
-                OrderItems::query()->create([
-                    'order_id' => $order->id,
-                    'product_id'=> $cartItem->product->id,
-                    'quantity' => $cartItem->quantity
-                ]);
-            }
+                    $cartItems = $user?->cart?->items->toArray();
 
-            $this->setDefaultDataForUser($user, 'phone', $request->validated('phone'));
-            $this->setDefaultDataForUser($user, 'default_address', $request->validated('address'));
+                    if (!$cartItems) return false;
 
+                    $order->items()->createMany($cartItems);
 
-            $user->cart()->delete();
-            Cache::delete('orders');
+                    $this->setDefaultDataForUser($user, 'phone', $request->validated('phone'));
+                    $this->setDefaultDataForUser($user, 'default_address', $request->validated('address'));
 
-            return response()->json([
-                'success' => true,
-                'message' => 'new order was created'
-            ], 201);
+                    $user->cart()->delete();
+                    Cache::delete('orders');
+
+                    return true;
+
+                } catch (\Exception $exception) {
+                    return false;
+                }
+
+            });
         }
 
         return response()->json([
-            'success' => false,
-            'message' => 'order has not been created'
-        ], 400);
+            'success' => $created,
+            'message' => $created ? 'success created' : 'has not been created'
+        ], $created ? 201 : 400);
 
     }
 
+
+    /**
+     * @param User|Authenticate $user
+     * @param string $key
+     * @param string $value
+     * @return void
+     * Установка дефолтных параметров для юзера
+     */
     private function setDefaultDataForUser(User|Authenticate $user, string $key, string $value): void
     {
         if (!$user->$key) {
@@ -91,18 +119,31 @@ class OrdersService
         }
     }
 
-    public function changeStatus(Order $order, int $statusId): JsonResponse
+    /**
+     * @param Order $order
+     * @param int $statusId
+     * @return OrderResource
+     * Изменить статус заказа по АПИ (Задел на микросервис)
+     */
+    public function changeStatus(Order $order, int $statusId): OrderResource
     {
         $status = Status::query()->find($statusId);
         if ($status) $order->update(['status_id' => $status->id]);
-        return response()->json(new OrderResource($order));
+        return new OrderResource($order);
     }
 
-    public function changePayment(Order $order, int $paymentId): JsonResponse
+
+    /**
+     * @param Order $order
+     * @param int $paymentId
+     * @return OrderResource
+     * Изменить тип оплаты по АПИ (Возможно для клиента)
+     */
+    public function changePayment(Order $order, int $paymentId): OrderResource
     {
         $payment = Payment::query()->find($paymentId);
         if ($payment) $order->update(['payment_id' => $payment->id]);
-        return response()->json(new OrderResource($order));
+        return new OrderResource($order);
     }
 
 }
